config/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-querycount==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
